Factor out `chat`
