+++ async-openai/src/types/mod.rs
@@ -4,6 +4,7 @@ mod assistant;
 mod assistant_file;
 mod assistant_impls;
 mod audio;
+mod chat;
 mod common;
 mod edit;
 mod embedding;
@@ -22,6 +23,7 @@ mod thread;
 pub use assistant::*;
 pub use assistant_file::*;
 pub use audio::*;
+pub use chat::*;
 pub use common::*;
 pub use edit::*;
 pub use embedding::*;
